Bug fix: jjmaestro https://github.com/jjmaestro found part of this bug
jjmaestro reported that central_authorizations didn't stop later checkings no matter
what they returned, which was right. He actually committed a bug fix for that issue.

However when reading his code I realized that the problem was bigger. I was forcing
the user to return a bool in central_authorizations, which was wrong. If the user
returns a bool, I pass it up, otherwise we keep checking rules.

Please update django-rules in your machine!