Redoing rules storage to be in memory
After a year without commits in the project, we are back trying to improve
performance and usability. Although the project didn't have any known issues,
we wanted to give it a facelift and enhance the storage, enhance how rules are
defined, clean up the internal naming conventions, etc.

From now on, rules will be stored in a dictionary in memory, not needing to
syncdb when installing the app. Above all, this adds portability and simplicity
while giving a speed boost (not necessarily much if you had your rules table
cached, etc).

A bit of the in-memory dictionary code is based on the django-rules fork
by Chris Glass. He is already listed as a contributor of this project.

With respect to how we deal with decorators, we now pass the model in a much
explicit way. This helps alleviate the "one unique codename in the app" while
ensuring that you are more explicitly defining the rule to check in the
decorator.

All tests are passing with this new storage and decorators.

As a final note, we plan the next version of django-rules to be backwards
incompatible, due to the huge refactors that will be done. In any case, we will
try to be as careful as possible to help everybody transition into the next
awesome version.